Extract chunkArray to its own file
