+++ frontend/fingertips-frontend/lib/ViewsHelpers.test.ts
@@ -1,4 +1,4 @@
-import { chunkArray, getHealthDataForIndicator } from './ViewsHelpers';
+import { getHealthDataForIndicator } from './ViewsHelpers';
 import {
   IndicatorsApi,
   IndicatorWithHealthDataForArea,
@@ -20,16 +20,6 @@ const mockIndicator: IndicatorWithHealthDataForArea = {
     },
   ],
 };
-
-describe('chunkArray', () => {
-  it('should chunk an array into the correct sized sub arrays', () => {
-    const testArray1: string[] = new Array(10).fill('a', 0, 10);
-    const testArray2: string[] = new Array(5).fill('b', 0, 5);
-
-    expect(chunkArray(testArray1, 2)).toHaveLength(5);
-    expect(chunkArray(testArray2, 2)).toEqual([['b', 'b'], ['b', 'b'], ['b']]);
-  });
-});
 
 describe('getHealthDataForIndicator', () => {
   const mockIndicatorsApi = mockDeep<IndicatorsApi>();
